requirements.txt
annotated-types==0.7.0
anyio==4.5.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pypdfium2==4.30.0
pytesseract==0.3.13
python-multipart==0.0.12
reportlab==4.2.5
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.31.1
